requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.7.2
    # via
    #   -r requirements.in
    #   django
build==1.0.3
    # via
    #   -r requirements.in
    #   pip-tools
click==8.1.7
    # via
    #   -r requirements.in
    #   pip-tools
django==5.0.2
    # via -r requirements.in
django-admin-extra-buttons==1.5.7
    # via -r requirements.in
dnspython==2.6.1
    # via -r requirements.in
flake8==7.0.0
    # via -r requirements.in
mccabe==0.7.0
    # via
    #   -r requirements.in
    #   flake8
packaging==23.2
    # via
    #   -r requirements.in
    #   build
pip-tools==7.3.0
    # via -r requirements.in
pycodestyle==2.11.1
    # via
    #   -r requirements.in
    #   flake8
pyflakes==3.2.0
    # via
    #   -r requirements.in
    #   flake8
pyproject-hooks==1.0.0
    # via
    #   -r requirements.in
    #   build
sqlparse==0.4.4
    # via
    #   -r requirements.in
    #   django
wheel==0.42.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
